STA_LSTM/requirements.txt
h5py==3.8.0
matplotlib==3.7.1
#mediapipe==0.9.0.1
#opencv-python==4.5.2.52
pandas==1.3.5
Pillow==9.5.0
PyQt5==5.15.9
pytz==2023.3
#scikit-learn==1.0.2
scipy==1.7.3
tensorflow==2.10.0
torch==1.13.1
torchvision==0.14.1
ttkbootstrap==1.10.1

numpy==1.24.3